@@ -1,19 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="FluentMigrator" Version="3.2.17" />
-    <PackageReference Include="FluentMigrator.Runner" Version="3.2.17" />    
-    <PackageReference Include="Dapper" Version="2.0.90" />
+    <PackageReference Include="FluentMigrator" Version="5.2.0" />
+    <PackageReference Include="FluentMigrator.Runner" Version="5.2.0" />    
+    <PackageReference Include="Dapper" Version="2.1.35" />
     <PackageReference Include="Dapper.FluentMap" Version="2.0.0" />
-    <PackageReference Include="Dommel" Version="2.3.2" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Dapper.Fluent.ORM.MultiSchema\Dapper.Fluent.ORM.MultiSchema.csproj" />
+    <ProjectReference Include="..\Dommel\Dommel.csproj" />
   </ItemGroup>
 
 </Project>
