@@ -5,10 +5,9 @@ using System.Text;
 using System.Threading.Tasks;
 using Dapper.Fluent.ORM.Contracts;
 
-namespace Dapper.Fluent.ORM.Postgres.Contracts
-{
-    public interface IPostgresRepository<TEntity> : IDapperRepository<TEntity> where TEntity : class
-    {
+namespace Dapper.Fluent.ORM.Postgres.Contracts;
 
-    }
+public interface IPostgresRepository<TEntity> : IDapperRepository<TEntity> where TEntity : class
+{
+
 }
