@@ -5,26 +5,28 @@ using Dapper.Fluent.ORM.Postgres.Contracts;
 using FluentMigrator.Runner;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Dapper.Fluent.ORM.Postgres.Extensions
+namespace Dapper.Fluent.ORM.Postgres.Extensions;
+
+public static class StartupExtensions
 {
-    public static class StartupExtensions
+    public static IServiceCollection AddPostgresRepositoryWithMigration(this IServiceCollection services, string connectionString, bool enablePooling, string defaultSchema = "public", params Assembly[] assembliesWithMappers)
     {
-        public static IServiceCollection AddPostgresRepositoryWithMigration(this IServiceCollection services, string connectionString, string defaultSchema = "public", params Assembly[] assembliesWithMappers)
+        services.AddScoped(typeof(IRepositorySettings), service => new PostgresRepositorySettings
         {
-            services.AddScoped(typeof(IRepositorySettings), service => new PostgresRepositorySettings
-            {
-                ConnString = connectionString,
-                DefaultSchema = defaultSchema ?? "public"
-            });
+            ConnString = connectionString,
+            DefaultSchema = defaultSchema ?? "public",
+            EnablePooling = enablePooling
+        });
 
-            services.AddScoped(typeof(IPostgresRepository<>), typeof(PostgresRepository<>));
+        services
+            .AddTransient(typeof(IPostgresRepository<>), typeof(PostgresRepository<>))
+            .AddScoped<IJsonPropertyHandler, PostgresJsonPropertyHandler>();
 
-            return services
-                .AddMigrator()
-                .ConfigureRunner(cfg => cfg
-                    .ConfigureMigrator(connectionString, assembliesWithMappers)
-                    .AddPostgres()
-                );
-        }
+        return services
+            .AddMigrator()
+            .ConfigureRunner(cfg => cfg
+                .ConfigureMigrator(connectionString, assembliesWithMappers)
+                .AddPostgres()
+            );
     }
 }
