@@ -1,10 +1,9 @@
 ﻿using Dapper.FluentMap.Mapping;
 
-namespace Dapper.Fluent.ORM.Mapping
+namespace Dapper.Fluent.ORM.Mapping;
+
+public interface IDapperFluentPropertyMap : IPropertyMap
 {
-    public interface IDapperFluentPropertyMap : IPropertyMap
-    {
-        object DefaultValue { get; }
-        int Lenght { get; }
-    }
+    object DefaultValue { get; }
+    int Lenght { get; }
 }
