@@ -1,5 +1,7 @@
 ﻿using System;
+using Dapper.Fluent.ORM.Contracts;
 using Dapper.Fluent.ORM.Mapping;
+using Dapper.FluentMap;
 using Dapper.FluentMap.Mapping;
 using Dommel;
 
@@ -7,23 +9,33 @@ namespace Dapper.Fluent.Mapping.Resolvers
 {
     public class TableNameResolver : ITableNameResolver
     {
+        private readonly ISchema _schema;
         private static readonly ITableNameResolver DefaultResolver = new DefaultTableNameResolver();
 
-        /// <inheritdoc />
+        public TableNameResolver(ISchema schema)
+        {
+            _schema = schema;
+        }
+
         public string ResolveTableName(Type type)
         {
-            IEntityMap entityMap;
-            if (FluentMap.FluentMapper.EntityMaps.TryGetValue(type, out entityMap))
-            {
-                var mapping = entityMap as IDapperFluentEntityMap;
+            if (!FluentMapper.EntityMaps.TryGetValue(type, out var entityMap))
+                return DefaultResolver.ResolveTableName(type);
 
-                if (mapping != null)
-                {
-                    return mapping.TableName;
-                }
+            if (!(entityMap is IDapperFluentEntityMap mapping))
+                return DefaultResolver.ResolveTableName(type);
+
+            var tableName = mapping.TableName;
+
+            var schema = _schema.GetSchema();
+
+            if (tableName.Contains(".") && mapping.IsDynamicSchema && !string.IsNullOrWhiteSpace(schema))
+            {
+                return $"{schema}.{tableName.Split(".")[1]}";
             }
 
-            return DefaultResolver.ResolveTableName(type);
+            return tableName;
+
         }
     }
 }
